Refactor externals/omniscidb to heavydb
